requirements.txt
# =============================================================
# Savarkar GPT - Dependencies
# =============================================================

# --- Core Framework ---
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-text-splitters>=0.3.0

# --- Embedding Model ---
sentence-transformers>=3.0.0
torch>=2.0.0

# --- Vector Database ---
pinecone-client>=5.0.0
langchain-pinecone>=0.2.0

# --- Google Gemini ---
google-generativeai>=0.8.0

# --- Web Server ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# --- Utilities ---
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# --- PDF Processing (already used) ---
PyMuPDF>=1.24.0
